billing_project/requirementes.txt
asgiref==3.6.0
Django==4.2
django-cors-headers==3.14.0
django-phonenumber-field==7.0.2
django-phonenumbers==1.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Faker==18.4.0
mysqlclient==2.1.1
phonenumbers==8.13.8
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.3
tzdata==2023.3
